requirements.txt
asgiref==3.7.2
Django==3.2.12
django-cors-headers==4.3.0
django-environ==0.11.2
django-model-utils==4.3.1
django-notifications-hq==1.8.3
django-role-permissions==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
inflection==0.5.1
jsonfield==3.1.0
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
swapper==1.3.0
uritemplate==4.1.1
